LayoutTests/imported/w3c/web-platform-tests/fetch/corb/style-css-mislabeled-as-html.sub-expected.txt

Blocked access to external URL http://www1.localhost:8800/fetch/corb/resources/css-mislabeled-as-html.css
Blocked access to external URL http://www1.localhost:8800/fetch/corb/resources/css-mislabeled-as-html.css
Header example

Paragraph body


PASS CSS is not applied (because of strict content-type enforcement for cross-origin stylesheets)
